refactor: extract call resolution strategy dispatch to resolver/strategy.ts
Extracts resolveByReceiver and resolveByGlobal from resolveByMethodOrGlobal
into src/domain/graph/resolver/strategy.ts. The dispatcher shrinks from
cognitive=107/bugs=1.48 to cognitive=4/bugs=0.14. isModuleScopedLanguage
moves to strategy.ts and is re-exported from call-resolver.ts for backward
compatibility with build-edges.ts.

docs check acknowledged
